+++ kbstatpy/__init__.py
@@ -1,5 +1,5 @@
 from .options import KbstatOptions
 from .kbstat import Kbstat
 
-__version__ = "1.14.0"
+__version__ = "1.14.1"
 __all__ = ["Kbstat", "KbstatOptions", "__version__"]